pyproject.toml
[build-system]
requires = [
    "setuptools >= 64",
    "setuptools_scm >= 6.4"
]
build-backend = "setuptools.build_meta"

[project]
name = "asphalt-serialization"
description = "Serialization component for the Asphalt framework"
readme = "README.rst"
authors = [{name = "Alex Grönholm", email = "alex.gronholm@nextday.fi"}]
license = {text = "Apache License 2.0"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "asphalt ~= 4.8"
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/asphalt-framework/asphalt-serialization"

[project.optional-dependencies]
msgpack = ["msgpack ~= 1.0"]
cbor = ["cbor2 ~= 5.0"]
yaml = ["ruamel.yaml >= 0.15"]
test = [
    "anyio[trio] >= 4.1",
    "asphalt-serialization[msgpack,cbor,yaml]",
    "coverage >= 7",
    "pytest >= 7",
]
doc = [
    "asphalt-serialization[msgpack,cbor,yaml]",
    "Sphinx >= 7",
    "sphinx_rtd_theme >= 1.3.0",
    "sphinx-autodoc-typehints >= 1.2.0",
]

[project.entry-points."asphalt.components"]
serialization = "asphalt.serialization:SerializationComponent"

[project.entry-points."asphalt.serialization.serializers"]
cbor = "asphalt.serialization.serializers.cbor:CBORSerializer"
json = "asphalt.serialization.serializers.json:JSONSerializer"
msgpack = "asphalt.serialization.serializers.msgpack:MsgpackSerializer"
pickle = "asphalt.serialization.serializers.pickle:PickleSerializer"
yaml = "asphalt.serialization.serializers.yaml:YAMLSerializer"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "dirty-tag"

[tool.ruff.lint]
extend-select = [
    "ASYNC",        # flake8-async
    "W",            # pycodestyle warnings
    "G",            # flake8-logging-format
    "I",            # isort
    "ISC",          # flake8-implicit-str-concat
    "PGH",          # pygrep-hooks
    "RUF",          # Ruff-specific rules
    "UP",           # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["asphalt.serialization"]

[tool.pytest.ini_options]
addopts = "-rsx --tb=short"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
strict = true
explicit_package_bases = true

[tool.coverage.run]
source = ["asphalt.serialization"]
relative_files = true
branch = true

[tool.coverage.report]
show_missing = true

[tool.tox]
env_list = ["py39", "py310", "py311", "py312", "py313", "pypy3"]
skip_missing_interpreters = true

[tool.tox.env_run_base]
commands = [["python", "-m", "pytest", { replace = "posargs", extend = true }]]
package = "editable"
extras = ["test"]

[tool.tox.env.pyright]
deps = ["pyright"]
commands = [["pyright", "--verifytypes", "asphalt.exceptions"]]

[tool.tox.env.docs]
commands = [["sphinx-build", "-W", "-n", "docs", "build/sphinx", { replace = "posargs", extend = true }]]
extras = ["doc"]
